Allow namespace in source/filter/kind
Previously the source, filter, or kind must exist on the root directory of each special directories.
With this commit, plugin users can create a package to group them source, filter, or kind for example

+- rplugin/python3/denite/
    +- source/
        +- __init__.py
        +- my_plugin/
            +- __init__.py (define 'Source')
            +- a.py
            +- b/
                +- __init__.py (empty)
                +- c.py
            +- d/
                +- __init__.py (define 'Source')
                +- e.py

And names of each sources will be

    my_plugin       (my_plugin/__init__.py)
    my_plugin/a     (my_plugin/a.py)
    my_plugin/b/c   (my_plugin/b/c.py)
    my_plugin/d     (my_plugin/d/__init__.py)
    my_plugin/d/e   (my_plugin/d/e.py)

Note that 'my_plugin/b' is not listeda above while I assumed that
'my_plugin/b/__init__.py' is an empty file.